# shiro-daemon/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "shiro API",
    "description": "Utoipa's axum bindings for seamless integration for the two",
    "contact": {
      "name": "Juha Kukkonen",
      "email": "juha7kukkonen@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/model/profiles/delete": {
      "post": {
        "tags": [
          "model"
        ],
        "summary": "删除档案；删除默认档案时默认项回退到剩余第一个",
        "operationId": "delete_profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "已删除（含更新后列表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileListResponse"
                }
              }
            }
          },
          "401": {
            "description": "未鉴权"
          },
          "404": {
            "description": "档案不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "写入失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/api/v1/model/profiles/list": {
      "post": {
        "tags": [
          "model"
        ],
        "summary": "档案列表（含默认档案指定）",
        "operationId": "list_profiles",
        "responses": {
          "200": {
            "description": "档案列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileListResponse"
                }
              }
            }
          },
          "401": {
            "description": "未鉴权"
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/api/v1/model/profiles/save": {
      "post": {
        "tags": [
          "model"
        ],
        "summary": "注册/更新档案（同 key 覆盖；首个档案自动成为默认）",
        "operationId": "upsert_profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "已保存（含更新后列表）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileListResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未鉴权"
          },
          "500": {
            "description": "写入失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
    "/api/v1/model/profiles/test": {
      "post": {
        "tags": [
          "model"
        ],
        "summary": "测试档案连接（发一个最小真实调用 ping；验证端点 + 密钥 + 模型 + 协议）",
        "operationId": "test_profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "测试结果（ok 标识连通与否，message 说明）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileTestResponse"
                }
              }
            }
          },
          "401": {
            "description": "未鉴权"
          },
          "404": {
            "description": "档案不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "ProfileKeyRequest": {
        "type": "object",
        "required": [
          "key"
        ],
        "properties": {
          "key": {
            "type": "string",
            "description": "档案 key"
          }
        }
      },
      "ProfileListResponse": {
        "type": "object",
        "required": [
          "profiles"
        ],
        "properties": {
          "default_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "默认档案 key（无档案时为 null）"
          },
          "profiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileSummary"
            }
          }
        }
      },
      "ProfileSummary": {
        "type": "object",
        "required": [
          "key",
          "base_url",
          "model",
          "protocol",
          "api_key_set"
        ],
        "properties": {
          "api_key_preview": {
            "type": [
              "string",
              "null"
            ]
          },
          "api_key_set": {
            "type": "boolean"
          },
          "base_url": {
            "type": "string"
          },
          "key": {
            "type": "string",
            "description": "档案 key（供应商 preset key）"
          },
          "model": {
            "type": "string",
            "description": "该档案当前使用的模型"
          },
          "protocol": {
            "type": "string",
            "description": "协议：openai | anthropic"
          },
          "thinking": {
            "type": [
              "string",
              "null"
            ],
            "description": "思考强度（未启用为 null）"
          }
        }
      },
      "ProfileTestResponse": {
        "type": "object",
        "required": [
          "ok",
          "message"
        ],
        "properties": {
          "latency_ms": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "测试耗时（毫秒，成功时返回）",
            "minimum": 0
          },
          "message": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "UpsertProfileRequest": {
        "type": "object",
        "required": [
          "key",
          "base_url",
          "model"
        ],
        "properties": {
          "api_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "新密钥；缺省或空字符串 = 保留原值（新建档案必填）"
          },
          "base_url": {
            "type": "string"
          },
          "key": {
            "type": "string",
            "description": "档案 key（供应商 preset key；同 key 覆盖更新）"
          },
          "model": {
            "type": "string"
          },
          "protocol": {
            "type": [
              "string",
              "null"
            ]
          },
          "thinking": {
            "type": [
              "string",
              "null"
            ],
            "description": "思考强度：缺省/空 = 不启用；low / medium / high"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}